+++ src/berconpy/errors.py
@@ -1,11 +1,20 @@
+import asyncio
+
+
 class RCONError(Exception):
     """The base class for RCON errors."""
 
 
 class LoginFailure(RCONError):
-    """Raised when either the RCON server could not respond to
-    login attempts, or the password given to the server was incorrect.
-    """
+    """Raised when the client could not log into the RCON server."""
+
+
+class LoginRefused(LoginFailure):
+    """Raised when the password given to the RCON server was incorrect."""
+
+
+class LoginTimeout(LoginFailure, asyncio.TimeoutError):
+    """Raised when the RCON server could not respond to our login attempts."""
 
 
 class RCONCommandError(RCONError):